Updates to PR
Thank you for taking the time to review and make changes. I've made some changes to it. The bigest one is to remove the two sets of details on authentication from the initial setup instructions. I'd rather keep people on track. If they want to follow up on options, they can dig into the 'Authentication' section.